requirements.txt
# Core ML and AI libraries
torch>=1.9.0
sentence-transformers>=2.2.0
transformers>=4.20.0
numpy>=1.21.0

# Vector database and similarity search
faiss-cpu>=1.7.0
# For GPU acceleration, use: faiss-gpu>=1.7.0

# PDF processing
pypdf>=3.0.0

# Web framework and API
flask>=2.0.0
requests>=2.25.0
waitress>=2.1.0

# Data processing and utilities
pandas>=1.3.0
scikit-learn>=1.0.0

# Optional: For better performance
# accelerate>=0.20.0
# optimum>=1.8.0

# Development and testing (optional)
# pytest>=6.0.0
# black>=22.0.0
# flake8>=4.0.0
